added a type cast to prevent the uint32 to float conversion warning

+++ src/Background.cpp
@@ -40,7 +40,7 @@ void Background::Draw(void) {
 			glBindTexture(GL_TEXTURE_2D, texture->textureID);
 			glUniform1i(glGetUniformLocation(*shader, "textureSampler"), 0);
 			// TODO: Change to not use SDL_Ticks, due to SDL_Ticks being consistent in its values
-			glUniform1f(glGetUniformLocation(*shader, "time"), SDL_GetTicks());	// Pass SDL_Ticks to the shader as a timer
+			glUniform1f(glGetUniformLocation(*shader, "time"), (float)SDL_GetTicks());	// Pass SDL_Ticks to the shader as a timer
 		}
 		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, currentMesh.indicesBufferObject);
 		glDrawElements(GL_TRIANGLES, (GLsizei)currentMesh.indices.size(), GL_UNSIGNED_INT, (void*)0);
